requirements.txt
absl-py==0.11.0
astor==0.8.1
astunparse==1.6.3
backcall==0.2.0
blis==0.7.4
cached-property==1.5.2
cachetools==4.2.0
catalogue==1.0.0
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
colorama==0.4.4
cycler==0.10.0
cymem==2.0.5
dataclasses==0.8
decorator==4.4.2
Flask==1.1.2
flatbuffers==1.12
gast==0.2.2
google-api-core==1.23.0
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-cloud-vision==2.0.0
google-pasta==0.2.0
googleapis-common-protos==1.52.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
importlib-metadata==3.3.0
imutils==0.5.3
ipython==7.16.1
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.10.3
joblib==1.0.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
libcst==0.3.15
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.1.1
murmurhash==1.0.5
mypy-extensions==0.4.3
nltk==3.5
numpy==1.19.3
oauthlib==3.1.0
opencv-python==4.4.0.46
opt-einsum==3.3.0
parso==0.7.1
pickleshare==0.7.5
Pillow==6.2.2
plac==1.1.3
preshed==3.0.5
prompt-toolkit==3.0.8
proto-plus==1.13.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.7.3
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.4
PyYAML==5.3.1
regex==2020.11.13
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.6
scipy==1.5.4
simplejson==3.17.2
six==1.15.0
srsly==1.0.5
tensorboard==2.0.2
tensorboard-plugin-wit==1.7.0
tensorflow-estimator==2.0.1
tensorflow-gpu==2.0.0
termcolor==1.1.0
thinc==7.4.5
tqdm==4.54.1
traitlets==4.3.3
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.26.2
wasabi==0.8.0
wcwidth==0.2.5
Werkzeug==0.16.1
wrapt==1.12.1
zipp==3.4.0
